Move yeepClient to admin_ui/app/yeepClient.js
