+++ admin_ui/app/yeepClient.js
@@ -1,7 +1,8 @@
-const YeepClient = require('@yeep/client');
+import YeepClient from '@yeep/client';
 
 const yeepClient = new YeepClient({
   baseURL: process.env.API_BASE_URL,
+  onError: (err) => console.error(err),
 });
 
 export default yeepClient;
